Simplify ToTvp as class-specific extension
